seccomp: build script for embedding the default filters
