--- src/vmm/src/default_syscalls/macros.rs
@@ -1,22 +1,0 @@
-// Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/// Shorthand for chaining `SeccompCondition`s with the `and` operator  in a `SeccompRule`.
-/// The rule will take the `Allow` action if _all_ the conditions are true.
-///
-/// [`Allow`]: enum.SeccompAction.html
-/// [`SeccompCondition`]: struct.SeccompCondition.html
-/// [`SeccompRule`]: struct.SeccompRule.html
-macro_rules! and {
-    ($($x:expr,)*) => (SeccompRule::new(vec![$($x),*], SeccompAction::Allow));
-    ($($x:expr),*) => (SeccompRule::new(vec![$($x),*], SeccompAction::Allow))
-}
-
-/// Shorthand for chaining `SeccompRule`s with the `or` operator in a `SeccompFilter`.
-///
-/// [`SeccompFilter`]: struct.SeccompFilter.html
-/// [`SeccompRule`]: struct.SeccompRule.html
-macro_rules! or {
-    ($($x:expr,)*) => (vec![$($x),*]);
-    ($($x:expr),*) => (vec![$($x),*])
-}
